pyproject.toml
[tool.poetry]
name = "realty_back"
version = "0.1.0"
description = ""
authors = ["dl <danil.spb.it@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
psycopg2-binary = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
